+++ BookingsApi/BookingsApi.DAL/Mappings/ScreeningMap.cs
@@ -1,5 +1,3 @@
-
-
 using BookingsApi.Domain.SpecialMeasure;
 
 namespace BookingsApi.DAL.Mappings;
@@ -23,19 +21,5 @@ public class ScreeningMap : IEntityTypeConfiguration<Screening>
         builder.HasOne(screening => screening.Endpoint)
             .WithOne(participant => participant.Screening)
             .HasForeignKey<Screening>(screening => screening.EndpointId);
-    }
-}
-
-public class ScreeningEntityMap : IEntityTypeConfiguration<ScreeningEntity>
-{
-    public void Configure(EntityTypeBuilder<ScreeningEntity> builder)
-    {
-        builder.ToTable(nameof(ScreeningEntity));
-
-        builder.HasKey(x => x.Id);
-
-        builder.HasOne(x => x.Screening)
-            .WithMany(x => x.ScreeningEntities)
-            .HasForeignKey(x => x.ScreeningId);
     }
 }
